PR 2.2: cog collaborator migration
Move COGMixin (~380 LoC) from src/pyramids/dataset/ops/cog.py:COGMixin
into src/pyramids/dataset/_collaborators.py:COG. Migrated members:

* to_cog (method) — Cloud Optimized GeoTIFF write with all GDAL COG
  driver options (compress, blocksize, tiling_scheme, target_srs, etc.)
* is_cog (property) — gdal_validate_cog wrapper, False for MEM/vsimem.
* validate_cog (method) — full ValidationReport, with strict mode.
* _warn_if_categorical_with_averaging (private) — guardrail emitting a
  UserWarning when an averaging resampler runs on a categorical raster.

Inside the collaborator, every read of Dataset state (_raster, file_name)
now goes through self._ds.<state>. The module-level constants
_AVERAGING_RESAMPLERS and _INTEGER_DTYPES move with the methods.

src/pyramids/dataset/dataset.py:

* Drop _COGMixin from the import alias list and from Dataset's MRO.
* Add three Dataset facades: to_cog, is_cog (property), validate_cog —
  each thin-forwards to self.cog.<X>. Existing call-sites continue to
  work via the facade: dataset/collection.py:1200 (slice_ds.to_cog) and
  dataset/ops/io.py:153 (ds.to_cog) need no change.

src/pyramids/dataset/ops/__init__.py: drop COGMixin import and __all__
entry. src/pyramids/dataset/ops/cog.py: deleted.

tests/dataset/test_collaborators.py:

* Move (cog, to_cog) and (cog, validate_cog) rows from FORWARDING_METHODS
  to FACADE_METHODS — covered by TestFacadeDelegation.
* Move (cog, is_cog) from READONLY_PROPERTIES to a new FACADE_PROPERTIES
  list, covered by a new test_facade_property_reads_from_collaborator
  test method that patches the collaborator-class property and asserts
  Dataset.is_cog reads from it.

No NetCDF override touches any cog method. Full main test suite passes;
only the two pre-existing Windows file-lock flakes in
test_dataset_collection.py remain failing.
